refactor/consistent model names
