+++ mealie/db/models/_all_models.py
@@ -1,5 +1,5 @@
-from db.models.meal_models import *
-from db.models.recipe_models import *
-from db.models.settings_models import *
-from db.models.theme_models import *
+from db.models.mealplan import *
+from db.models.recipe import *
+from db.models.settings import *
+from db.models.theme import *
 from db.models.users import *
