+++ mealie/db/models/settings.py
@@ -1,7 +1,6 @@
 import sqlalchemy as sa
 import sqlalchemy.orm as orm
 from db.models.model_base import BaseMixins, SqlAlchemyBase
-from db.models.recipe_models import Category
 
 
 class SiteSettingsModel(SqlAlchemyBase, BaseMixins):
